api shop complete

+++ routes/api.php
@@ -18,13 +18,13 @@ use Illuminate\Support\Facades\Route;
 //    return $request->user();
 //});
 
-
-Route::get('/users', [\App\Http\Controllers\UserController::class, 'index'])->middleware('auth:sanctum');
+//users
+Route::get('/users', [\App\Http\Controllers\UserController::class, 'index'])->middleware(['auth:sanctum','admin']);
 Route::prefix('/user')->group(function () {
     Route::post('/create', [\App\Http\Controllers\UserController::class, 'store']);
-    Route::get('/index/{index}', [\App\Http\Controllers\UserController::class, 'show'])->middleware('auth:sanctum');
-    Route::put('/update/{id}',[\App\Http\Controllers\UserController::class,'update'])->middleware('auth:sanctum');
-    Route::delete('/delete/{id}',[\App\Http\Controllers\UserController::class,'destroy'])->middleware('auth:sanctum');
+    Route::get('/index/{index}', [\App\Http\Controllers\UserController::class, 'show'])->middleware(['auth:sanctum','admin']);
+    Route::put('/update/{id}',[\App\Http\Controllers\UserController::class,'update'])->middleware(['auth:sanctum','admin']);
+    Route::delete('/delete/{id}',[\App\Http\Controllers\UserController::class,'destroy'])->middleware(['auth:sanctum','admin']);
     Route::post('/login', [\App\Http\Controllers\UserController::class, 'login']);
     Route::get('/logout', [\App\Http\Controllers\UserController::class, 'logout'])->middleware('auth:sanctum');
     Route::post('/info', [\App\Http\Controllers\UserController::class, 'userInfo'])->middleware('auth:sanctum');
@@ -33,21 +33,35 @@ Route::prefix('/user')->group(function () {
 
 
 //province
-Route::apiResource('province', \App\Http\Controllers\ProvinceController::class);
+Route::apiResource('province', \App\Http\Controllers\ProvinceController::class)->middleware(['auth:sanctum','admin'])->only(['store','update','destroy']);
 Route::get('/provinces/{province}/children', [\App\Http\Controllers\ProvinceController::class, 'cities']);
+Route::get('/province',[\App\Http\Controllers\ProvinceController::class,'index']);
+Route::get('/province/{id}',[\App\Http\Controllers\ProvinceController::class,'show']);
 
-Route::apiResource('city', \App\Http\Controllers\CityController::class);
+//cites
+Route::apiResource('city', \App\Http\Controllers\CityController::class)->middleware(['auth:sanctum','admin'])->only(['store','update','destroy']);
+Route::get('/city',[\App\Http\Controllers\CityController::class,'index']);
+Route::get('/city/{id}',[\App\Http\Controllers\CityController::class,'show']);
 
-
-Route::apiResource('brands', \App\Http\Controllers\BrandController::class)->middleware('auth:sanctum');
+//brands
+Route::apiResource('brand', \App\Http\Controllers\BrandController::class)->middleware(['auth:sanctum','admin'])->only(['store','update','destroy']);
 Route::get('/brands/{brand}/products', [\App\Http\Controllers\BrandController::class, 'products'])->middleware('auth:sanctum');
-Route::apiResource('categories', \App\Http\Controllers\CategoryController::class)->middleware('auth:sanctum');
+Route::get('/brand',[\App\Http\Controllers\BrandController::class,'index']);
+Route::get('/brand/{id}',[\App\Http\Controllers\BrandController::class,'show']);
+
+//categories
+Route::apiResource('category', \App\Http\Controllers\CategoryController::class)->middleware(['auth:sanctum','admin'])->only(['store','update','destroy']);
 Route::get('/categories/{category}/children', [\App\Http\Controllers\CategoryController::class, 'children'])->middleware('auth:sanctum');
 Route::get('/categories/{category}/parent', [\App\Http\Controllers\CategoryController::class, 'parent'])->middleware('auth:sanctum');
-
 Route::get('/categories/{category}/products', [\App\Http\Controllers\CategoryController::class, 'products'])->middleware('auth:sanctum');
+Route::get('/category',[\App\Http\Controllers\CategoryController::class,'index']);
+Route::get('/category/{id}',[\App\Http\Controllers\CategoryController::class,'show']);
 
-Route::apiResource('products', \App\Http\Controllers\ProductController::class);
+//products
+Route::apiResource('products', \App\Http\Controllers\ProductController::class)->middleware('auth:sanctum')->only(['store','update','destroy']);
+Route::get('/products',[\App\Http\Controllers\ProductController::class,'index']);
+Route::get('/products/{id}',[\App\Http\Controllers\ProductController::class,'show']);
+
 
 //pay
 Route::post('/payment/send', [\App\Http\Controllers\PaymentController::class, 'send'])->middleware('auth:sanctum');
